requirements.txt
asgiref
certifi
crispy-bootstrap5
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
Markdown
MouseInfo
mss
numpy
opencv-contrib-python
pillow
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
rubicon-objc
setuptools
sqlparse
typing_extensions
wheel
